+++ src/Bridge/VertexAi/CHANGELOG.md
@@ -6,6 +6,9 @@ CHANGELOG
 
  * Add support for Gemini 3.1 Flash Lite preview model (`gemini-3.1-flash-lite-preview`)
  * Add support for Gemini 3 Flash preview model (`gemini-3-flash-preview`)
+ * [BC BREAK] `ResultConverter` now returns a `MultiPartResult` when there are multiple `parts` in a `candidate`
+ * [BC BREAK] `ResultConverter` now returns `ExecutableCodeResult` and `CodeExecutionResult` parts when using `code_execution` server tool
+ * [BC BREAK] Throwing when code execution server tool fails is replaced with `CodeExecutionResult::isSucceeded()`
 
 0.7
 ---